requirements.txt
fastapi
uvicorn
Pillow
torchvision
python-multipart